Move encryptor to it's own file
