+++ internal/encryption/encryption.go
@@ -1,11 +1,7 @@
 package encryption
 
 import (
-	"github.com/opentofu/opentofu/internal/encryption/method"
 	"github.com/opentofu/opentofu/internal/encryption/registry"
-
-	"github.com/hashicorp/hcl/v2"
-	"github.com/zclconf/go-cty/cty"
 )
 
 // Encryption contains the methods for obtaining a StateEncryption or PlanEncryption correctly configured for a specific
@@ -25,21 +21,6 @@ type encryption struct {
 	reg registry.Registry
 }
 
-type encryptor struct {
-	cfg *Config
-	reg registry.Registry
-
-	// Used to evaluate hcl expressions
-	ctx *hcl.EvalContext
-
-	metadata map[string][]byte
-
-	// Used to build EvalContext (and related mappings)
-	keyValues    map[string]map[string]cty.Value
-	methodValues map[string]map[string]cty.Value
-	methods      map[string]method.Method
-}
-
 // New creates a new Encryption instance from the given configuration and registry.
 func New(reg registry.Registry, cfg *Config) Encryption {
 	return &encryption{
@@ -47,36 +28,6 @@ func New(reg registry.Registry, cfg *Config) Encryption {
 		reg: reg,
 	}
 }
-func (e *encryption) newEncryptor(meta map[string][]byte) (*encryptor, hcl.Diagnostics) {
-	var diags hcl.Diagnostics
-
-	enc := &encryptor{
-		cfg: e.cfg,
-		reg: e.reg,
-
-		ctx: &hcl.EvalContext{
-			Variables: map[string]cty.Value{},
-		},
-
-		metadata: meta,
-
-		keyValues:    make(map[string]map[string]cty.Value),
-		methodValues: make(map[string]map[string]cty.Value),
-		methods:      make(map[string]method.Method),
-	}
-
-	diags = append(diags, enc.setupKeyProviders()...)
-	if diags.HasErrors() {
-		return nil, diags
-	}
-	diags = append(diags, enc.setupMethods()...)
-	if diags.HasErrors() {
-		return nil, diags
-	}
-
-	return enc, diags
-}
-
 func (e *encryption) StateFile() StateEncryption {
 	return NewState(e, e.cfg.StateFile, "statefile")
 }
@@ -88,7 +39,6 @@ func (e *encryption) PlanFile() PlanEncryption {
 func (e *encryption) Backend() StateEncryption {
 	return NewState(e, e.cfg.Backend, "backend")
 }
-
 func (e *encryption) RemoteState(name string) StateEncryption {
 	for _, remoteTarget := range e.cfg.Remote.Targets {
 		if remoteTarget.Name == name {
